requirements.txt
streamlit 
sentence_transformers
torch 
farm-haystack
